Bug fixes and improved UI visibility
Fixed crucial bug that reset the entire runHistory if the user mistyped username/password or left it out when restarting an unfinished run. Fixed a bug that allowed the user to restart a run without accepting the required pages. Added indication to view-receipt page when treatment provider does not have a practitioner ID (ydernummer). Additionally, total grant amount is now shown if there are multiple grants given for a single citizen.