Fixed issues with resuming runs
Fixed an issue that did not allow the user to resume a run in which they had confirmed processing all receipts with actions but no rejected receipts.
Fixed an issue where the correct procedure for calculating grants were not followed for some specific cases (type B when the facilitator was not state certified).